Using WHERE RN = 1 in SQL.
`MostTransactionsPerYearByEmployee.sql`: Retrieves the age, first name, last name, number of transactions, and age of the employee who conducted the most transactions in each year. This SQL code retrieves the age, first name, last name, number of transactions, and age of the employee who conducted the most transactions in each year. It first calculates the count of orders per employee per year, along with their first name, last name, and age. Then, it uses the ROW_NUMBER() window function to rank the employees based on the number of transactions in descending order within each year. Finally, it filters the results to select only the top-ranked employee in each year.